开发自动投递和黑白名单功能

+++ job/runner.py
@@ -1,18 +1,15 @@
-from dataBase import connection, build
+from dataBase import build
 from detail import job
 from detail import job_Detial
-from itools import job_AverWage, job_if, cooky, send_email
+from itools import job_AverWage, cooky, send_email, job_apply
 import datetime
 import codecs
 import csv
 from maps import coordinate, distance
 import socket
 import os
+from datafilling import data_Filling, white_List, black_List
 
-
-conn = connection()
-cur = conn.cursor()
-cur.execute("USE job_CD")
 timeout = 20
 socket.setdefaulttimeout(timeout)  # 这里对整个socket层设置超时时间。后续文件中如果再使用到socket，不必再设置
 
@@ -42,38 +39,9 @@ def run(jobarea, homeAddress, homeCity, email, income, subject, *args):
     distance(homeAddress, homeCity)
     print('工作直线距离计算完毕')
 
-    date_time = []
-    for i in range(2):
-        a = datetime.date.today() - datetime.timedelta(days=i)
-        a = str(a)
-        date_time.append(a)
-    date_time = tuple(date_time)
-    print(date_time)
-    cur.execute("DROP TABLE if exists job_Detail")
-    cur.execute("create table job_Detail (select w.job_Name,w.job_Wage,w.job_AverWage,w.company_Name,w.company_Nature,w.company_Scale,w.company_Address,c.company_Distance,c.company_Duration,c.company_Traffic,w.job_PeopleNum,w.job_Issue,w.job_Article,w.job_Link from company c left join work w on c.company_Id=w.company_Id )")
-    cur.execute(
-        "select job_Name,job_Wage,job_AverWage,company_Name,company_Nature,company_Scale,company_Address,company_Distance,company_Duration,company_Traffic,job_PeopleNum,job_Issue,left(job_Article,300),job_Link from job_Detail where (company_Duration<=2500 or company_Duration='') and (job_Issue in {0})".format(date_time))
-    result = cur.fetchall()
-    cur.execute(
-        "select COLUMN_NAME from INFORMATION_SCHEMA.Columns where table_name='job_Detail' and table_schema='job_cd'")
-    title = cur.fetchall()
-# 将结果中已经投递过的工作记录去掉
-    result = list(result)
-    print("总记录数：", len(result))
-    cookies = cooky('cookies1.txt')
-    print(cookies)
-    headers = {
-        'Host': 'm.51job.com',
-        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/56.0.2924.87 Safari/537.36',
-        'Accept': 'text/css,*/*;q=0.1',
-        'Connection': 'keep-alive'
-    }
-    for i in result:
-        print(i[13])
-        if job_if(i[13], headers, cookies) == True:
-            result.remove(i)
-            print("成功移除一条已申请的记录", i[13])
-    print("去除已经重复投递后的工作数量", len(result))
+    # 进行数据整理
+    result, title = data_Filling()
+
     # 生成结果CSV文件
     with codecs.open("job_Detail.csv", "w", encoding="utf_8_sig") as f:
         f_csv = csv.writer(f)
@@ -98,8 +66,51 @@ def run(jobarea, homeAddress, homeCity, email, income, subject, *args):
         else:
             print("邮件发送成功")
             break
-    cur.close()
-    conn.close()
+# 执行数据清理工作
+    whitelist = white_List()
+    blacklist = black_List()
+    result = list(result)
+    filling = []
+    for i in result:
+        for j in whitelist:
+            if j in i[0]:
+                filling.append(i)
+                print(i[0], j)
+                break
+    for i in filling:
+        for j in blacklist:
+            if i[15] == j:
+                filling.remove(i)
+# 将筛选后的结果发送到邮箱
+    with codecs.open("job_Detail.csv", "w", encoding="utf_8_sig") as f:
+        f_csv = csv.writer(f)
+        f_csv.writerow(title)
+        f_csv.writerows(filling)
+        print("自动投递文件生成")
+    while True:
+        try:
+            send_email('smtp.qq.com', '272861776@qq.com', 'xjsdroroibjacaej',
+                       'louse12345@163.com', "自动投递简历的列表", '清理后的工作清单')
+        except Exception as e:
+            print('发生邮件错误，错误原因为:', e)
+        else:
+            print("邮件发送成功")
+            break
+# 自动投递简历功能
+    a = input("是否需要进行自动投递？(Y/N)")
+    if a == 'Y':
+        result = list(result)
+        headers = {
+            'Host': 'm.51job.com',
+            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/56.0.2924.87 Safari/537.36',
+            'Accept': 'text/css,*/*;q=0.1',
+            'Connection': 'keep-alive'
+        }
+        cookies = cooky('cookies1.txt')
+        for i in result:
+            print(i[13], i[14])
+            job_apply(i[13], headers, cookies, i[14])
+            print("已投递企业信息：", i[0], i[3])
 
 if __name__ == "__main__":
     jobarea = '090200'  # 提供基本参数，广东030000，四川090000，省会编码是0200
@@ -107,11 +118,15 @@ if __name__ == "__main__":
     keyword2 = "运营"
     keyword3 = "品牌"
     homeAddress = '锦江区东风路4号一栋一单元'
+    whitelist = []
+    blacklist = []
     homeCity = "成都"
     email = 'larkjoe@126.com'
     income = int('5000')
     subject = "宝宝鸡-{0}的工作记录，请查收".format(datetime.date.today())
+    close = input("是否需要执行自动关机功能？(Y/N)")
     build()
     run(jobarea, homeAddress, homeCity, email,
         income, subject, keyword1, keyword2, keyword3)
-    os.system('shutdown -s -t 200')
+    if close == "Y":
+        os.system('shutdown -s -t 200')
